Ensured that all comments were up2d8 and fixed small bug
When the board would fill up without a winner, the ai would still continue playing, which is obviously invalid. So I added a condition for the ai playing; that being that the board cannot be full ;) Ya know, normal logic stuff. Silly that SUCH A SILLY BUG MADE IT IN